Add User and Event services
- Add services to implement User and Event business logic
- Add DTO for object creation
- Update models and add migrations

+++ Infrastructure/Repositories/User/IUserRepository.cs
@@ -4,7 +4,7 @@ namespace EventPlanner.Repository
 {
     public interface IUserRepository : IRepository<User>
     {
-        public Task<User> GetUserByTelegramIdAsync(int telegramId);
+        public Task<User> GetUserByTelegramIdAsync(long telegramId);
     }
 
 }
